+++ libgcompat/ctype.c
@@ -1,0 +1,1 @@
+const unsigned short int *__ctype_b;