+++ libgcompat/internal.c
@@ -1,7 +1,6 @@
-#include <stdio.h>
 #include <stdarg.h>
+#include <stdio.h>
 #include <stdlib.h>
-
 
 void GCOMPAT__panic(const char *fmt, ...)
 {